requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-mail-templated==2.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
